Servicio Envio terminado y sus respectivos Test

+++ src/main/java/co/edu/uniquindio/unimarket/servicios/implementacion/EnvioServicioImpl.java
@@ -1,6 +1,7 @@
 package co.edu.uniquindio.unimarket.servicios.implementacion;
 
 import co.edu.uniquindio.unimarket.dto.EnvioDTO;
+import co.edu.uniquindio.unimarket.dto.EnvioGetDTO;
 import co.edu.uniquindio.unimarket.entidades.Compra;
 import co.edu.uniquindio.unimarket.entidades.Envio;
 import co.edu.uniquindio.unimarket.repositorios.CompraRepo;
@@ -8,6 +9,8 @@ import co.edu.uniquindio.unimarket.repositorios.EnvioRepo;
 import co.edu.uniquindio.unimarket.servicios.interfaces.EnvioServicio;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 @AllArgsConstructor
@@ -17,65 +20,102 @@ public class EnvioServicioImpl implements EnvioServicio {
     private final CompraRepo compraRepo;
 
     @Override
-    public Envio crearEnvio(EnvioDTO envioDTO, int idCompra) throws Exception {
+    public EnvioGetDTO crearEnvio(EnvioDTO envioDTO, int idCompra) throws Exception {
+
+        // Verificar si el id de la compra es válido
+        if (!compraRepo.existsById(idCompra))
+            throw new Exception("No se encontró la compra con el id suministrado");
 
         // Recuperar la compra de la base de datos
-        Compra compra = compraRepo.findById(idCompra)
-                .orElseThrow(() -> new Exception("No se encontró la compra con el id especificado"));
+        Compra compra = compraRepo.findById(idCompra).get();
 
-        Envio envio;
+        Envio envio = convertir(envioDTO);
 
         if (compra.getEnvio() != null) {
             // Actualizar el envío existente con los nuevos datos
             envio = compra.getEnvio();
-        } else {
-            // Crear un nuevo envío
-            envio = new Envio();
             envio.setDireccionEnvio(envioDTO.getDireccionEnvio());
             envio.setCiudadEnvio(envioDTO.getCiudadEnvio());
             envio.setTelefono(envioDTO.getTelefono());
             envio.setFechaEntregaEstimada(envioDTO.getFechaEntregaEstimada());
-
+        } else {
             // Establecer la relación entre el envío y la compra
             compra.setEnvio(envio);
             envio.setCompra(compra);
-
-            envioRepo.save(envio);
         }
 
-        return envio;
+        envioRepo.save(envio);
+
+        return convertir(envio);
     }
 
     @Override
-    public Envio actualizarEnvio(EnvioDTO envioDTO, int idEnvio) throws Exception {
+    public EnvioGetDTO actualizarEnvio(int idEnvio, EnvioDTO envioDTO) throws Exception {
+
+        // Verificar si el id del envío es válido
+        if (!envioRepo.existsById(idEnvio))
+            throw new Exception("No se encontró el envío con el id suministrado");
 
         // Recuperar el envío de la base de datos
-        Envio envio = envioRepo.findById(idEnvio)
-                .orElseThrow(() -> new Exception("No se encontró el envío con el id especificado"));
+        Envio envio = convertir(envioDTO);
 
-        // Actualizar los atributos del envío
+        envio.setIdEnvio(idEnvio);
+
+        return convertir(envioRepo.save(envio));
+    }
+
+
+    @Override
+    public int eliminarEnvio(int idEnvio) throws Exception {
+
+        // Verificar si el id del envío es válido
+        if (!envioRepo.existsById(idEnvio))
+            throw new Exception("No se encontró el envío con el id suministrado");
+
+        // Recuperar el envío de la base de datos
+        Envio envio = envioRepo.findById(idEnvio).get();
+
+        // Eliminar el envío
+        envioRepo.delete(envio);
+
+        return idEnvio;
+    }
+
+    @Override
+    public EnvioGetDTO obtenerEnvio(int idEnvio) throws Exception {
+        return convertir(obtener(idEnvio));
+    }
+
+
+    public Envio obtener(int idEnvio) throws Exception {
+        Optional<Envio> envio = envioRepo.findById(idEnvio);
+
+        if (envio.isEmpty()) {
+            throw new Exception("el codigo" + idEnvio + "no esta asociado a ningun envio");
+        }
+
+        return envio.get();
+    }
+
+    private EnvioGetDTO convertir(Envio envio) {
+        EnvioGetDTO envioGetDTO = new EnvioGetDTO();
+        envioGetDTO.setIdEnvio(envio.getIdEnvio());
+        envioGetDTO.setDireccionEnvio(envio.getDireccionEnvio());
+        envioGetDTO.setCiudadEnvio(envio.getCiudadEnvio());
+        envioGetDTO.setTelefono(envio.getTelefono());
+        envioGetDTO.setFechaEntregaEstimada(envio.getFechaEntregaEstimada());
+        return envioGetDTO;
+    }
+
+    private Envio convertir(EnvioDTO envioDTO) throws Exception {
+
+        Envio envio = new Envio();
         envio.setDireccionEnvio(envioDTO.getDireccionEnvio());
         envio.setCiudadEnvio(envioDTO.getCiudadEnvio());
         envio.setTelefono(envioDTO.getTelefono());
         envio.setFechaEntregaEstimada(envioDTO.getFechaEntregaEstimada());
 
-        // Guardar los cambios en la base de datos
-        envioRepo.save(envio);
-
         return envio;
-    }
-
-
-    @Override
-    public void eliminarEnvio(int idEnvio) throws Exception {
-
-        // Recuperar el envío de la base de datos
-        Envio envio = envioRepo.findById(idEnvio)
-                .orElseThrow(() -> new Exception("No se encontró el envío con el id especificado"));
-
-        // Eliminar el envío
-        envioRepo.delete(envio);
-
     }
 }
